Add support for local files mirror as buildroot primary download site
Buildroot uses wget to download source code but wget can't access local files directly so
we need to setup a local web server that wget can talk to.

Add http://localhost:8000/PATH_TO_LOCAL_FILES as Primary download site in buildroot and
mount that path as a volume in docker.